Intermediate/Crypter/Backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic[email]==2.7.1
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
requests==2.31.0
bcrypt